refactor(collection): use kysely expression builder for category usage query
Replaces raw SQL OR clause with kysely's eb.or() in
getCategoryOptionUsageCount, tightens pageId schema to integers, and
adds a test covering the folder-parent (non-collection) case.

+++ apps/studio/src/schemas/collection.ts
@@ -94,7 +94,7 @@ export const getCollectionTagsSchema = z
 
 export const getCategoryOptionUsageCountSchema = z.object({
   siteId: z.number().min(1),
-  pageId: z.number().min(1),
+  pageId: z.number().int().min(1),
   categoryId: z.string().uuid(),
 })
